refactor: 섹션별 폴더 구조 정리 및 파일 재배치 #33
