+++ src/components/sections/DetailSection/DetailHero/DetailHero.css
