+++ src/components/sections/DetailSection/DetailInfo/DetailInfo.css
