+++ src/components/sections/DetailSection/DetailMedia/DetailMedia.css
