+++ src/components/sections/DetailSection/DetailPeople/DetailPeople.css
