+++ src/components/sections/DetailSection/DetailPeople/PeopleSwiper.jsx
