+++ src/components/sections/DetailSection/DetailPeople/PersonCard.jsx
