+++ src/pages/MovieDetail.jsx
@@ -3,11 +3,11 @@ import { useParams } from "react-router-dom";
 import { getBestTrailerUrl } from "../api/videos";
 import { useMovieDetailFull } from "../hooks/useMovieDetailFull";
 import { sortCrew } from "../utils/sort";
-import DetailHero from "../components/sections/DetailSection/DetailHero";
+import DetailHero from "../components/sections/DetailSection/DetailHero/DetailHero";
 import TrailerModal from "../components/sections/NowPlayingSection/TrailerModal";
-import DetailPeople from "../components/sections/DetailSection/DetailPeople";
-import DetailMedia from "../components/sections/DetailSection/DetailMedia";
-import DetailInfo from "../components/sections/DetailSection/DetailInfo";
+import DetailPeople from "../components/sections/DetailSection/DetailPeople/DetailPeople";
+import DetailMedia from "../components/sections/DetailSection/DetailMedia/DetailMedia";
+import DetailInfo from "../components/sections/DetailSection/DetailInfo/DetailInfo";
 
 const MovieDetail = () => {
   const { id } = useParams();
